Corrected comments: - moved CustomButton; - deleted unnecessary code; - launching UserProfileActivity using implicit intent; - added more information about user.
