+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,12 @@
         tools:targetApi="31">
         <activity
             android:name=".view.details.UserProfileActivity"
-            android:exported="false" />
+            android:exported="false">
+            <intent-filter>
+                <action android:name="show user" />
+                <category android:name="android.intent.category.DEFAULT" />
+            </intent-filter>
+        </activity>
         <activity
             android:name=".view.list.MainActivity"
             android:exported="true">
