+++ app/src/main/java/by/yancheuski/githubusers/data/UsersRepoImpl.kt
@@ -1,12 +1,9 @@
 package by.yancheuski.githubusers.data
 
-import android.os.Handler
-import android.os.Looper
 import by.yancheuski.githubusers.domain.entities.UserEntity
+import by.yancheuski.githubusers.domain.entities.UserProfileEntity
 import by.yancheuski.githubusers.domain.repos.UsersRepo
 import io.reactivex.rxjava3.core.Single
-
-private const val DATA_LOADING_FAKE_DELAY = 3_000L
 
 class UsersRepoImpl : UsersRepo {
 
@@ -16,23 +13,9 @@ class UsersRepoImpl : UsersRepo {
         UserEntity("pjhyett", 3, "https://avatars.githubusercontent.com/u/3?v=4"),
     )
 
-    override fun getUsers(onSuccess: (List<UserEntity>) -> Unit, onError: ((Throwable) -> Unit)?) {
-        Handler(Looper.getMainLooper()).postDelayed({
-            onSuccess(data)
-        }, DATA_LOADING_FAKE_DELAY)
-    }
-
     override fun getUsers(): Single<List<UserEntity>> = Single.just(data)
 
-    override fun getUser(
-        login: String,
-        onSuccess: (user: UserEntity) -> Unit,
-        onError: ((Throwable) -> Unit)?,
-    ) {
-        TODO("Not yet implemented")
-    }
-
-    override fun getUser(login: String): Single<UserEntity> {
+    override fun getUser(login: String): Single<UserProfileEntity> {
         TODO("Not yet implemented")
     }
 }
