+++ app/src/main/java/by/yancheuski/githubusers/data/retrofit/GitHubUsersApi.kt
@@ -1,6 +1,7 @@
 package by.yancheuski.githubusers.data.retrofit
 
 import by.yancheuski.githubusers.domain.entities.UserEntity
+import by.yancheuski.githubusers.domain.entities.UserProfileEntity
 import io.reactivex.rxjava3.core.Single
 import retrofit2.http.GET
 import retrofit2.http.Path
@@ -11,5 +12,5 @@ interface GitHubUsersApi {
     fun getListOfUsers(): Single<List<UserEntity>>
 
     @GET("users/{login}")
-    fun getUserGit(@Path("login") login: String): Single<UserEntity>
+    fun getUserGit(@Path("login") login: String): Single<UserProfileEntity>
 }
