+++ app/src/main/java/by/yancheuski/githubusers/view/list/MainActivity.kt
@@ -11,7 +11,6 @@ import by.yancheuski.githubusers.databinding.ActivityMainBinding
 import by.yancheuski.githubusers.domain.entities.UserEntity
 import by.yancheuski.githubusers.view.OnClickUserListener
 import by.yancheuski.githubusers.view.details.USER_EXTRA_KEY
-import by.yancheuski.githubusers.view.details.UserProfileActivity
 import io.reactivex.rxjava3.disposables.CompositeDisposable
 
 class MainActivity : AppCompatActivity(), OnClickUserListener {
@@ -34,14 +33,7 @@ class MainActivity : AppCompatActivity(), OnClickUserListener {
     }
 
     private fun showListOfUsers() {
-
-        // Class CustomButton
         binding.loadButton.buttonObservable.subscribe { viewModel.loadData() }
-
-        // Extension function
-//        binding.loadButton.observableClickListener().subscribe {
-//            viewModel.loadData()
-//        }
     }
 
     private fun initViewModel() {
@@ -83,7 +75,7 @@ class MainActivity : AppCompatActivity(), OnClickUserListener {
     }
 
     override fun onClickUser(user: UserEntity) {
-        startActivity(Intent(this@MainActivity, UserProfileActivity::class.java)
+        startActivity(Intent("show user")
             .apply {
                 putExtra(USER_EXTRA_KEY, user.login)
             })
