+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     tools:context=".view.list.MainActivity">
 
-    <by.yancheuski.githubusers.utils.CustomButton
+    <by.yancheuski.githubusers.view.components.CustomButton
         android:id="@+id/load_button"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
