+++ app/src/main/res/values/dimens.xml
@@ -13,6 +13,10 @@
     <dimen name="size_margin_top">36dp</dimen>
     <dimen name="text_size_login_profile_text_view">42sp</dimen>
     <dimen name="text_size_id_profile_text_view">30sp</dimen>
-    <dimen name="size_margin_top_id">24dp</dimen>
+    <dimen name="size_margin_top_main">24dp</dimen>
     <dimen name="margin_default_button">10dp</dimen>
+    <dimen name="text_size_title_text_view">16sp</dimen>
+    <dimen name="margin_start_default_text_view">18dp</dimen>
+    <dimen name="margin_top_default_text_view">12dp</dimen>
+    <dimen name="margin_start_info_text_view">16dp</dimen>
 </resources>
